+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Features:
 * Add `Brightness` module support
 * Add `Battery` module support for FreeBSD
 * Add `--disk-show-unknown` option for Disk module
+* Add `--gpu-hide-integrated` option (#379)
+* Add `--gpu-hide-discrete` option (#379)
 
 Logos:
 * Raspbian (@IamNoRobot, #373)